+++ packages/zen-mapper/src/zen_mapper/__init__.py
@@ -1,10 +1,16 @@
 import logging
 
 from .adapters import sk_learn, to_networkx
-from .cover import Width_Balanced_Cover, precomputed_cover, rectangular_cover
+from .cover import (
+    Data_Balanced_Cover,
+    Width_Balanced_Cover,
+    precomputed_cover,
+    rectangular_cover,
+)
 from .mapper import mapper
 
 __all__ = [
+    "Data_Balanced_Cover",
     "Width_Balanced_Cover",
     "mapper",
     "precomputed_cover",